+++ jdaviz/tests/test_show.py
@@ -1,17 +1,11 @@
-import sidecar
+import pytest
 
 
-def test_show_inline(specviz_helper):
-    res = specviz_helper.show_inline()
-
-    assert res is None
-
-
-def test_show_sidecar(specviz_helper):
-    res = specviz_helper.show_in_sidecar()
-    assert isinstance(res, sidecar.Sidecar)
+def test_show_popout(specviz_helper):
+    with pytest.raises(RuntimeError, match='Error in displaying Jdaviz'):
+        specviz_helper.show(loc="popout")
 
 
-def test_show_new_tab(specviz_helper):
-    res = specviz_helper.show_in_new_tab()
-    assert isinstance(res, sidecar.Sidecar)
+def test_show_invalid_mode(specviz_helper):
+    with pytest.raises(RuntimeError, match='Error in displaying Jdaviz'):
+        specviz_helper.show(loc="thisisnotavaliddisplaymode")