+++ CHANGES.rst
@@ -50,6 +50,9 @@ API Changes
   have been renamed to "Colormap" and "Monochromatic," respectively, for all image
   viewers. [#1406]
 
+- Viz tool display changed to ``viz.show()`` from ``viz.app``. Sidecar no longer returned by 
+  show methods. [#965]
+
 Cubeviz
 ^^^^^^^